getPlaylist impl

+++ qm-music-parent/qm-music-core/src/main/java/com/github/chenqimiao/dto/PlaylistItemDTO.java
@@ -17,7 +17,7 @@ public class PlaylistItemDTO {
 
     private Long songId;
 
-    private Long sortOrder;
+    private Integer sortOrder;
 
     private Long gmtCreate;
